AlexRestaurant

+++ src/main/java/com/umss/dev/training/jtemplate/persistence/repository/RegisterRestaurantRepository.java
@@ -5,6 +5,6 @@ import com.umss.dev.training.jtemplate.persistence.domain.RegisterRestaurant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface RegisterRestaurantRepository extends JpaRepository<RegisterRestaurant, Integer> {
+public interface RegisterRestaurantRepository extends JpaRepository<RegisterRestaurant, Long> {
 
 }
